+++ app/components/NoteView.vue
@@ -1,7 +1,7 @@
 <template>
   <div
     data-testid="note-view"
-    class="flex min-h-0 flex-1 flex-col overflow-hidden border border-border"
+    class="flex min-h-0 min-w-0 flex-1 flex-col border-t border-border"
   >
     <NoteTemplate />
   </div>